+++ src/main/java/duke/customClass/LogicHandler.java
@@ -1,10 +1,15 @@
-package customClass;
+package duke.customClass;
+
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.Todo;
 
 import java.util.List;
 
 public class LogicHandler {
     public void list(List<Task> list) {
-        // Check if the command is list and display the list of tasks.
+        // Check if the command is list and display the list of duke.tasks.
         String temp = "";
 
         for (int i = 0; i < list.size(); i++) {
@@ -38,7 +43,7 @@ public class LogicHandler {
             Todo todo = new Todo(input.split(" ", 2)[1]);
             list.add(todo);
             System.out.println("added: " + todo);
-            System.out.println("Now you have " + list.size() + " tasks in the list.");
+            System.out.println("Now you have " + list.size() + " duke.tasks in the list.");
         } catch (ArrayIndexOutOfBoundsException e) {
             System.out.println("Oops, your todo requires at least a description.");
         }
@@ -52,7 +57,7 @@ public class LogicHandler {
             Deadline deadline = new Deadline(task, date);
             list.add(deadline);
             System.out.println("added: " + deadline);
-            System.out.println("Now you have " + list.size() + " tasks in the list.");
+            System.out.println("Now you have " + list.size() + " duke.tasks in the list.");
         } catch (ArrayIndexOutOfBoundsException e) {
             System.out.println("Oops, your deadline requires both a description and date.");
         }
@@ -66,7 +71,7 @@ public class LogicHandler {
             Event event = new Event(task, date);
             list.add(event);
             System.out.println("added: " + event);
-            System.out.println("Now you have " + list.size() + " tasks in the list.");
+            System.out.println("Now you have " + list.size() + " duke.tasks in the list.");
         } catch (ArrayIndexOutOfBoundsException e) {
             System.out.println("Oops, your Event requires both a description and date");
         }
@@ -76,7 +81,7 @@ public class LogicHandler {
         try {
             int index = Integer.valueOf(input.split(" ", 2)[1]);
             System.out.println("Noted. I've removed the task:\n" + list.get(index - 1) +
-                    "\nNow you have " + (list.size() - 1) + " tasks in the list.");
+                    "\nNow you have " + (list.size() - 1) + " duke.tasks in the list.");
             list.remove(index - 1);
         } catch (ArrayIndexOutOfBoundsException e) {
             System.out.println("Delete requires a number");
