+++ src/main/java/duke/customClass/Parser.java
@@ -1,4 +1,4 @@
-package customClass;
+package duke.customClass;
 
 
 public class Parser {
