+++ src/main/java/duke/customClass/TaskList.java
@@ -1,6 +1,10 @@
-package customClass;
+package duke.customClass;
 
-import java.lang.reflect.Array;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.Todo;
+
 import java.util.List;
 import java.util.ArrayList;
