+++ src/main/java/duke/tasks/Task.java
@@ -1,4 +1,4 @@
-package customClass;
+package duke.tasks;
 
 public class Task {
     protected String description;
